Fix validation table formatting; reuse validationissues for config validation
